Revert "Try to use -pipe and -save-temps for gcc in ruby-head"
This reverts commit 89125cd1cc7588ba833007bbe249db2037b28270.

Using -pipe and -save-temps avoids issues with unicode characters in the TMP environment variable, but unfortunately it is not suitable to be used at all times.
When gcc is used to compile and link in one step, then -pipe (from CFLAGS) and -save-temps (from LDFLAGS) is combined in one command, which outputs a warning about ignoring -pipe when -save-temps is
used.
This warning is interpret as an error by build scripts like ruby's.

IMHO the narrow use case to work around unicode characters in TMP isn't worth the risk to break the build of other people.