Add a placeholder XML manifest to libruby all extension DLLs
This allows to add a customized manifest within the installer build process later on.
The Microsoft mt.exe allows to add a manifest to a EXE or DLL file after the build.
But there is no such tool in gcc/llvm.
So adding a placeholder manifest allows to later change the XML with simple text replacement.
That way the ruby build doesn't need any patches and is not bound too much to the installer.

The exe manifest of ruby core is re-enbaled now, since DLL manifests are sufficient to declare DLL dependencies.

Also use a build directory like the MSYS2/MINGW-packages.