Add a placeholder XML manifest to ruby.exe and all extension DLLs
This allows to add a customized manifest within the installer build process later on.
The Microsoft mt.exe allows to add a manifest to a EXE or DLL file after the build.
But there is no such tool in gcc/llvm.
So adding a placeholder manifest allows to later change the XML with simple text replacement.
That way the ruby build doesn't need any patches and is not bound too much to the installer.

Also use a build directory like the MSYS2/MINGW-packages.